style:경매페이지 반응형 UI
