+++ src/components/auctions/detail/AuctionInfo.tsx
@@ -1,14 +1,13 @@
 import { AuctionItem } from "@/_types/auctions/AuctionItem";
-import React, { useState } from "react";
+import React from "react";
 import { TimerStatus } from "./AuctionContent";
 import CustomIcon from "@/Icons/Icon";
-import CommonInput from "@/components/common/input/CommonInput";
 import Timer from "@/components/common/Timer";
 import dayjs from "dayjs";
 import BidUserInfo from "./BidUserInfo";
-import Button from "@/components/common/Button";
-import BidButton from "./BidButton";
 import { useGetSubscribeAuctionId } from "@/hooks/fetcher/auctions/useGetSubscribeAuctionId";
+import BidControl from "./BidControl";
+import useDeviceSize from "@/hooks/responsive/useDeviceSize";
 
 interface AuctionInfoProps {
   data: AuctionItem;
@@ -26,10 +25,9 @@ const AuctionInfo = ({
   auctionId,
 }: AuctionInfoProps) => {
   const { auction } = useGetSubscribeAuctionId(Number(auctionId));
+  const { isTablet } = useDeviceSize();
 
   const currentPrice = auction?.currentPrice || data.currentPrice;
-  const [bidPrice, setBidPrice] = useState<number>(currentPrice);
-  const [buyNowPrice, setBuyNowPrice] = useState<number | null>(null);
 
   return (
     <>
@@ -38,9 +36,11 @@ const AuctionInfo = ({
       </p>
       <div className="flex flex-col gap-0.25">
         <Timer startTime={start} endTime={end} status={status} />
-        <p className="text-2.5 font-bold tracking-[-0.8px]">
-          {currentPrice.toLocaleString()}원{" "}
-        </p>
+        {!isTablet && (
+          <p className="text-2.5 font-bold tracking-[-0.8px]">
+            {currentPrice.toLocaleString()}원{" "}
+          </p>
+        )}
       </div>
       <div className="flex flex-col gap-l-1">
         <div className="flex justify-between w-full">
@@ -69,63 +69,30 @@ const AuctionInfo = ({
           <p className="text-fgGrayDefault">{data.auctionCode}</p>
         </div>
       </div>
-      {status === "progress" && (
-        <div className="flex gap-0.75">
-          <div className="flex-1 w-full flex flex-col gap-l-0.25">
-            <p className="text-fgGrayDefault text-[14px]">즉시 구매가</p>
-            <div className="flex flex-col gap-l-0.5">
-              <CommonInput
-                placeholder={data.buyNowPrice?.toLocaleString() || ""}
-                value={buyNowPrice ? buyNowPrice.toLocaleString() : ""}
-                onChange={(e) => {
-                  const value = e.target.value.replace(/[^0-9]/g, "");
-                  setBuyNowPrice(value ? Number(value) : null);
-                }}
-                onBlur={(e) => {
-                  if (buyNowPrice) {
-                    e.target.value = buyNowPrice.toLocaleString();
-                  }
-                }}
-                disabled={!data.buyNowPrice}
-              />
-              <Button
-                title="즉시 구매하기"
-                variant="secondary"
-                width="100%"
-                disabled={!data.buyNowPrice}
-              />
-            </div>
-          </div>
-          <div className="flex-1 flex flex-col gap-l-0.25">
-            <p className="text-fgGrayDefault text-[14px]">희망 입찰가</p>
-            <div className="flex flex-col gap-l-0.5">
-              <button className="input-base input-default flex items-center justify-between ">
-                <div
-                  onClick={() =>
-                    setBidPrice((prev) => Math.max(0, prev - 10000))
-                  }
-                >
-                  <CustomIcon
-                    icon="CIRCLE-MINUS"
-                    className="w-[20px] h-[20px]"
-                  />
-                </div>
-                <div className="w-full cursor-default text-1 leading-[1.4] tracking-[-0.32px]">
-                  {bidPrice?.toLocaleString() || 0}
-                </div>
-                <div onClick={() => setBidPrice((prev) => prev + 10000)}>
-                  <CustomIcon
-                    icon="CIRCLE-PLUS"
-                    className="w-[20px] h-[20px]"
-                  />
-                </div>
-              </button>
-              <BidButton bidAmount={bidPrice} auctionId={Number(auctionId)} />
-            </div>
-          </div>
+      {!isTablet && (
+        <>
+          {status === "progress" && (
+            <BidControl
+              data={data}
+              currentPrice={currentPrice}
+              auctionId={auctionId}
+            />
+          )}
+          <BidUserInfo />
+        </>
+      )}
+      {isTablet && (
+        <div className="fixed bottom-0 w-full z-20 left-0 bg-bgGrayDepth2 pt-[12px] pb-2 px-[40px] border-t border-borderDivider flex flex-col gap-1.5">
+          <p className="text-1.75 font-bold tracking-[-0.8px]">
+            {currentPrice.toLocaleString()}원{" "}
+          </p>
+          <BidControl
+            data={data}
+            currentPrice={currentPrice}
+            auctionId={auctionId}
+          />
         </div>
       )}
-      <BidUserInfo />
     </>
   );
 };
